filter offsets are real numbers
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@8271 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/camd/camd.h
@@ -853,7 +853,7 @@ class Camera:public Rts2ScriptDevice
 
 	protected:
 		Rts2ValueSelection *camFilterVal;
-		rts2core::IntegerArray *camFilterOffsets;
+		rts2core::DoubleArray *camFilterOffsets;
 	
 	private:
 		Rts2ValueInteger *camFocVal;
